requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
Flask
Flask-SQLAlchemy
pymysql==1.0.2
Flask-Migrate==4.0.7
Pillow
pytest==7.4.2
pytest-cov==4.0.0
python-dotenv
flask-swagger-ui
